Tab 데이터 싱크 맞추기 및 네이밍 규칙 수정
