+++ scripts/initialize.js
@@ -1,9 +1,10 @@
-import { addEventToRotatingArrow } from "./addEventToRotatingArrow.js";
+
 import { addEventToRotatingBox } from "./addEventToRotatingBox.js";
 import { renderDefaultSceen } from "./render.js";
 import { loadSubscribeCompanies } from "./subscribe.js";
 import { addEventToToggle } from "./toggle.js";
 import { setUpCurrentTime } from "./util.js";
+import { addEventToRotatingArrow } from "./swiper.js";
 
 export function initialize(state) {
     setUpCurrentTime();
