+++ scripts/magicNumber.js
